+++ packages/bond-farming/contracts/interfaces/IRouter.sol
@@ -1,5 +1,5 @@
 pragma solidity >=0.8.0;
 
 interface IRouter {
-  function getAmountsOut(uint amountIn, address[] memory path) external view returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] memory path) external view returns (uint256[] memory amounts);
 }
