+++ packages/dAssets-liquidator/contracts/interfaces/IDYToken.sol
@@ -2,19 +2,27 @@
 pragma solidity >=0.8.9;
 
 interface IDYToken {
-  function deposit(uint _amount, address _toVault) external;
+    function deposit(uint256 _amount, address _toVault) external;
 
-  function depositTo(address _to, uint _amount, address _toVault) external;
+    function depositTo(
+        address _to,
+        uint256 _amount,
+        address _toVault
+    ) external;
 
-  function depositCoin(address to, address _toVault) external payable;
+    function depositCoin(address to, address _toVault) external payable;
 
-  function withdraw(address _to, uint _shares, bool needWETH) external;
+    function withdraw(
+        address _to,
+        uint256 _shares,
+        bool needWETH
+    ) external;
 
-  function underlyingTotal() external view returns (uint);
+    function underlyingTotal() external view returns (uint256);
 
-  function underlying() external view returns (address);
+    function underlying() external view returns (address);
 
-  function balanceOfUnderlying(address _user) external view returns (uint);
+    function balanceOfUnderlying(address _user) external view returns (uint256);
 
-  function underlyingAmount(uint amount) external view returns (uint);
+    function underlyingAmount(uint256 amount) external view returns (uint256);
 }
