+++ packages/over-collateralization/contracts/Constants.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.9;
 
 contract Constants {
-  uint internal constant PercentBase = 10000;
+    uint256 internal constant PercentBase = 10000;
 }
