+++ packages/over-collateralization/contracts/interfaces/IFeeConf.sol
@@ -3,5 +3,5 @@
 pragma solidity >=0.8.0;
 
 interface IFeeConf {
-  function getConfig(bytes32 _key) external view returns (address, uint); 
+    function getConfig(bytes32 _key) external view returns (address, uint256);
 }
