+++ packages/over-collateralization/contracts/interfaces/ILiquidateCallee.sol
@@ -2,6 +2,17 @@
 pragma solidity >=0.8.0;
 
 interface ILiquidateCallee {
-  function liquidateDeposit(address borrower, address underlying, uint amount, bytes calldata data) external;
-  function liquidateBorrow(address borrower, address underlying, uint amount, bytes calldata data) external;
+    function liquidateDeposit(
+        address borrower,
+        address underlying,
+        uint256 amount,
+        bytes calldata data
+    ) external;
+
+    function liquidateBorrow(
+        address borrower,
+        address underlying,
+        uint256 amount,
+        bytes calldata data
+    ) external;
 }
