+++ packages/over-collateralization/contracts/vault/BTCMintVault.sol
@@ -5,10 +5,8 @@ pragma solidity 0.8.9;
 import "./MintVault.sol";
 
 contract BTCMintVault is MintVault {
-
-// TODO: 不同的dAsset 可以可能有不同的实现mint。
-  function underlyingBurn(uint amount) internal virtual override {
-    IDUSD(underlying).burnme(amount);
-  }
-
+    // TODO: 不同的dAsset 可以可能有不同的实现mint。
+    function underlyingBurn(uint256 amount) internal virtual override {
+        IDUSD(underlying).burnme(amount);
+    }
 }
