+++ packages/stable-coin-minter/contracts/interfaces/TokenRecipient.sol
@@ -1,10 +1,10 @@
 pragma solidity >=0.6.0;
 
 interface TokenRecipient {
-  // must return ture
-  function tokensReceived(
-      address from,
-      uint amount,
-      bytes calldata exData
-  ) external returns (bool);
+    // must return ture
+    function tokensReceived(
+        address from,
+        uint256 amount,
+        bytes calldata exData
+    ) external returns (bool);
 }
